+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-### next
+<a name="v0.2.2"></a>
+### v0.2.2 - 2021-09-29
 - `-o` option to immediately prompt for password for drawer opening
 - ctrl-v pastes the content of the clipboard (if safecloset is compiled with "clipbpoard" feature)
+- mouse wheel support
 
 <a name="v0.2.1"></a>
 ### v0.2.1 - 2021-09-26